+++ NEWS.md
@@ -1,3 +1,29 @@
+# ggfittext 0.8.0
+
+## Major changes
+
+- New `geom_bar_text()` convenience function for labelling of bars in bar plots
+- New `contrast` argument for automatically contrasting a background fill
+  colour
+- New `fullheight` argument for including or excluding descenders in text size
+- Faster for many common uses
+- New `hjust` and `vjust` arguments that can be controlled independent of
+  `place`
+- New `outside` argument to allow text that doesn't fit inside the box to be
+  teleported outside of it 
+
+## Minor changes
+
+- `padding.x` and `padding.y` now both default to 1 mm
+
+## Bug fixes
+
+- When reflowing text, calculation of the best aspect ratio now takes into
+  account the output device's aspect ratio
+- When reflowing text, calculation of the best aspect ratio no longer refuses
+  to entertain the possibility that the original text might have the best ratio
+- Angled text is now placed correctly
+
 # ggfittext 0.7.0
 
 ## Major changes